backend/requirements.txt
tensorflow
numpy<2
opencv-python
matplotlib
pillow
kagglehub
fastapi
uvicorn
pymongo